in-progress jacobian evaluation
