+++ src/Model/PhasorDynamics/BusFault/BusFaultDependencyTracking.cpp
@@ -1,0 +1,11 @@
+#include "BusFaultImpl.hpp"
+
+namespace GridKit
+{
+  namespace PhasorDynamics
+  {
+    // Additional template instantiations
+    template class BusFault<DependencyTracking::Variable, long int>;
+    template class BusFault<DependencyTracking::Variable, size_t>;
+  }
+}
