+++ src/Model/PhasorDynamics/BusFault/CMakeLists.txt
@@ -4,5 +4,11 @@ gridkit_add_library(phasor_dynamics_bus_fault
   OUTPUT_NAME
     gridkit_phasor_dynamics_bus_fault)
 
+gridkit_add_library(phasor_dynamics_bus_fault_dependency_tracking
+  SOURCES
+    BusFaultDependencyTracking.cpp
+  OUTPUT_NAME
+    gridkit_phasor_dynamics_bus_fault_dependency_tracking)
+
 # Link to interface target for all components
 target_link_libraries(gridkit_phasor_dynamics_components INTERFACE GRIDKIT::phasor_dynamics_bus_fault)
