+++ src/Model/PhasorDynamics/Load/LoadImpl.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <cmath>
 #include <iostream>
