+++ src/Model/PhasorDynamics/SynchronousMachine/GENROUwS/CMakeLists.txt
@@ -4,5 +4,11 @@ gridkit_add_library(phasor_dynamics_genrou
   OUTPUT_NAME
     gridkit_phasor_dynamics_genrou)
 
+gridkit_add_library(phasor_dynamics_genrou_dependency_tracking
+  SOURCES
+    GenrouDependencyTracking.cpp
+  OUTPUT_NAME
+    gridkit_phasor_dynamics_genrou_dependency_tracking)
+
 # Link to interface target for all components
 target_link_libraries(gridkit_phasor_dynamics_components INTERFACE GRIDKIT::phasor_dynamics_genrou)
