+++ tests/UnitTests/PhasorDynamics/runSystemTests.cpp
@@ -9,6 +9,7 @@ int main()
   GridKit::Testing::SystemTests<double, size_t> test;
 
   result += test.constructor();
+  result += test.dependencyTracking();
   result += test.composer();
 
   return result.summary();
